cleanup, linting, minor tweaks
