+++ ui/src/app/cohort-search/cohort-builder/cohort-builder.component.html
@@ -33,7 +33,7 @@
             <div class="col-xs-8"><h3 class="mt-1">Include Participants</h3></div>
             <div class="col-xs-4 text-right">
               <button type="button" class="btn btn-outline" 
-                                    (click)="actions.initGroup('include')">
+                                    (click)="initGroup('include')">
                 <clr-icon shape="plus"></clr-icon> Add Group</button>
             </div>
           </div>
@@ -56,10 +56,9 @@
 
         <div class="col-xs-4">
           <div class="row">
-            <div class="col-xs-8"><h3 class="mt-1">Exclude Participants</h3></div>
-            <div class="col-xs-4 text-right">
+            <div class="col-xs-8"><h3 class="mt-1">Exclude Participants</h3></div> <div class="col-xs-4 text-right">
               <button type="button" class="btn btn-outline" 
-                                    (click)="actions.initGroup('exclude')">
+                                    (click)="initGroup('exclude')">
                 <clr-icon shape="plus"></clr-icon> Add Group</button>
             </div>
           </div>
