+++ ui/src/app/cohort-search/criteria-tree/index.ts
@@ -1,0 +1,3 @@
+export {CriteriaTreeRootComponent} from './root.component';
+export {CriteriaTreeNodeComponent} from './node.component';
+export {NodeInfoComponent} from './node-info.component';
