+++ ui/src/app/cohort-search/criteria-tree/node-info.component.ts
@@ -1,4 +1,9 @@
-import {Component, Input} from '@angular/core';
+import {
+  Component,
+  ChangeDetectionStrategy,
+  Input,
+  ViewEncapsulation
+} from '@angular/core';
 import {CohortSearchActions} from '../actions';
 import {Criteria} from 'generated';
 
@@ -6,6 +11,8 @@ import {Criteria} from 'generated';
 @Component({
   selector: 'app-node-info',
   templateUrl: './node-info.component.html',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  encapsulation: ViewEncapsulation.None,
 })
 export class NodeInfoComponent {
   @Input() node: Criteria;
