+++ ui/src/app/cohort-search/criteria-tree/root.component.html
@@ -6,7 +6,7 @@
 </ng-template>
 
 <ng-template #nodesLoaded>
-  <clr-tree-node *ngFor="let node of children">
+  <clr-tree-node *ngFor="let node of children; trackBy: trackById">
     <app-node-info [node]="node"></app-node-info>
 
     <ng-template clrIfExpanded>
