+++ ui/src/app/cohort-search/criteria-tree/root.component.ts
@@ -1,16 +1,21 @@
-import {Component, OnDestroy, OnInit, Input, ViewEncapsulation} from '@angular/core';
-import {NgRedux, select, dispatch} from '@angular-redux/store';
+import {
+  Component,
+  OnDestroy,
+  OnInit,
+  Input,
+  ViewEncapsulation,
+} from '@angular/core';
+import {NgRedux} from '@angular-redux/store';
 import {Subscription} from 'rxjs/Subscription';
-import 'rxjs/add/operator/mergeMap';
 
 import {CohortSearchActions} from '../actions';
 import {CohortSearchState} from '../store';
-import {CohortBuilderService, Criteria, SearchParameter} from 'generated';
 
 
 @Component({
   selector: 'app-tree-root',
   templateUrl: './root.component.html',
+  encapsulation: ViewEncapsulation.None,
 })
 export class CriteriaTreeRootComponent implements OnInit, OnDestroy {
 
@@ -39,5 +44,9 @@ export class CriteriaTreeRootComponent implements OnInit, OnDestroy {
 
   ngOnDestroy() {
     this.subscriptions.forEach(s => s.unsubscribe());
+  }
+
+  trackById(index, node) {
+    return node ? node.id : undefined;
   }
 }
