+++ ui/src/app/cohort-search/search-group-item/search-group-item.component.html
